# APIGOLANGMAP/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica o utilizador e gera o token para os próximos acessos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Realizar autenticação",
                "parameters": [
                    {
                        "description": "Do login",
                        "name": "evaluation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserLoginSwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Claims"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Desautentica o utilizador invalidando o token atual",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Realizar desautenticação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "bool"
                    },
                    "406": {
                        "description": "Cannot log out"
                    }
                }
            }
        },
        "/auth/refresh_token": {
            "put": {
                "description": "Atualiza o token de autenticação do usuário invalidando o antigo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Atualiza token de autenticação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Claims"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Cannot invalidate old token"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Regista um utilizador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Realizar registro",
                "parameters": [
                    {
                        "description": "Do register",
                        "name": "evaluation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserRegisterSwagger"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Claims"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/follower/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exibe a lista, sem todos os campos, de todos os followers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "summary": "Obtem os Followers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Follower"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Associa um Follower a um User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "summary": "Associa um Follower(User) a um User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Associate User as Follower",
                        "name": "follower",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.FollowingSwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Follower"
                            }
                        }
                    },
                    "400": {
                        "description": "Token Malformed, Check Syntax, Follower User ID Already Associated"
                    },
                    "404": {
                        "description": "User Not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Desassocia um Follower de um User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "summary": "Desassocia um Follower(User) de um User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deassociate Follower from User",
                        "name": "follower",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.FollowingSwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deassociation Success"
                    },
                    "400": {
                        "description": "Token Malformed, Check Syntax"
                    },
                    "404": {
                        "description": "User Not found"
                    }
                }
            }
        },
        "/follower/following": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exibe a lista, sem todos os campos, de todos os users que estamos a seguir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "summary": "Obtem os users que estamos a seguir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Follower"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/follower/history": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtem as localizações de um determinado Follower",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "summary": "Obtem as localizações de um Follower",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "get follower locations history",
                        "name": "follower",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.FollowingHistorySwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Position"
                            }
                        }
                    },
                    "400": {
                        "description": "Token Malformed, Check Syntax, Dates Malformed"
                    },
                    "401": {
                        "description": "User not authorized to check given User locations."
                    },
                    "404": {
                        "description": "User ID Not Found, "
                    }
                }
            }
        },
...
        "/user/alert/time/": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza a periodicidade de alerta determinando o tempo máximo até dar uma pessoa como perdida",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Atualiza a periodicidade de alerta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Udpdate Alert",
                        "name": "evaluation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UserPeriodUpdateSwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "406": {
                        "description": "Not acceptable"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Claims": {
            "type": "object",
            "properties": {
                "access_mode": {
                    "type": "integer"
                },
                "userid": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Follower": {
            "type": "object",
            "properties": {
                "FollowerUserID": {
                    "type": "integer"
                },
                "UserID": {
                    "type": "integer"
                }
            }
        },
        "model.Position": {
            "type": "object",
            "required": [
                "Latitude",
                "Longitude"
            ],
            "properties": {
                "Latitude": {
                    "type": "number"
                },
                "Longitude": {
                    "type": "number"
                },
                "UserId": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "access_mode": {
                    "type": "integer"
                },
                "alertTime": {
                    "type": "integer"
                },
                "sos": {
                    "type": "boolean"
                },
                "userFriends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Follower"
                    }
                },
                "userPositions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Position"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "swaggermodel.FollowingHistorySwagger": {
            "type": "object",
            "required": [
                "end",
                "followerID",
                "start"
            ],
            "properties": {
                "end": {
                    "type": "string",
                    "example": "2022-07-01"
                },
                "followerID": {
                    "type": "integer"
                },
                "start": {
                    "type": "string",
                    "example": "2022-01-01"
                }
            }
        },
        "swaggermodel.FollowingSwagger": {
            "type": "object",
            "properties": {
                "FollowerUserID": {
                    "type": "integer"
                }
            }
        },
...
        "swaggermodel.UserLoginSwagger": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "swaggermodel.UserPeriodUpdateSwagger": {
            "type": "object",
            "properties": {
                "alertTimeInHours": {
                    "type": "integer"
                }
            }
        },
        "swaggermodel.UserRegisterSwagger": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}